backend/requirements.txt
python-dotenv==0.7.1
Flask==0.12.2
Flask-Cors==3.0.3
RPi.GPIO==0.6.3
twython==3.6.0
kafka==1.3.5
